(refactor) ARCH-012: split AccountsService god-class along its seams
Extract three collaborating services composed by AccountsService:
PortfolioAnalyticsService (pure distribution math, no IO imports, 16
unit tests), GatewayWalletService (wallet CRUD, gateway balances and
immediate pricing, owns the require-gateway guard and balance_entry
helper) and PerpetualTradingService (leverage, position mode and
positions via an injected connector provider). AccountsService keeps
balance polling, state coordination and DB persistence and exposes
thin delegators with identical signatures, so routers, deps.py,
main.py and executor_service are unchanged. accounts_service.py goes
from 1887 to 1344 lines. Distribution outputs verified byte-identical
against the previous implementation.
